% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{compare_condition}
\alias{compare_condition}
\title{compare between two conditions}
\usage{
compare_condition(
  seurat_object,
  gsva,
  sample_column,
  condition_column,
  condition1,
  condition2
)
}
\arguments{
\item{seurat_object}{anotated seurat object}

\item{gsva}{GSVA results}

\item{sample_column}{Sample column}

\item{condition_column}{Condition column}

\item{condition1}{First condition to compare}

\item{condition2}{Second condition to compare}
}
\value{
Differentially expressed modules
}
\description{
compare between two conditions
}
