% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{do_dem}
\alias{do_dem}
\title{compare between two conditions}
\usage{
do_dem(
  seurat_object,
  sample_column = "batch",
  gmt = module_info,
  condition_column = "condition",
  condition1 = "SjS",
  condition2 = "HC"
)
}
\arguments{
\item{seurat_object}{anotated seurat object}

\item{sample_column}{Sample column}

\item{condition_column}{Condition column}

\item{condition1}{First condition to compare}

\item{condition2}{Second condition to compare}
}
\value{
Differentially expressed modules
}
\description{
compare between two conditions
}
