% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{draw_volcano}
\alias{draw_volcano}
\title{compare between two conditions}
\usage{
draw_volcano(comparison, p_cutoff = 0.05, logFC_cutoff = 0.25)
}
\arguments{
\item{results}{DEM results}
}
\value{
Differentially expressed modules
}
\description{
compare between two conditions
}
