% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{gene_enrichment_scimmuneco}
\alias{gene_enrichment_scimmuneco}
\title{compare between two conditions}
\usage{
gene_enrichment_scimmuneco(gene_list)
}
\arguments{
\item{gene_list}{INPUT GENE LIST}
}
\value{
gene_res Differentially expressed modules
}
\description{
compare between two conditions
}
