% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{gsva_cell_type}
\alias{gsva_cell_type}
\title{GSVA for one cell type}
\usage{
gsva_cell_type(
  seurat_object,
  cell_column = "predicted.celltype.l1",
  sample_column = "GEO_ident",
  condition_column = "Groups",
  cell,
  gmt = module_info
)
}
\arguments{
\item{seurat_object}{anotated seurat object}

\item{cell_column}{Cell type column}

\item{sample_column}{Sample column}

\item{condition_column}{Condition column}

\item{cell}{Cell type to subset}
}
\value{
A Seurat object with GSVA results
}
\description{
GSVA for one cell type
}
