% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{module_info}
\alias{module_info}
\title{Example Dataset from CSV}
\format{
A data frame with columns:
\describe{
\item{mpg}{Miles per gallon}
\item{cyl}{Number of cylinders}
...
}
}
\source{
Originally from a CSV file.
}
\usage{
module_info
}
\description{
A dataset imported from a CSV file.
}
\keyword{datasets}
